requirements.txt
streamlit
opencv-python
ultralytics
numpy
moviepy==1.0.3
edge-tts
Pillow
